added mag file
