+++ MicroMouse-2016.vcxproj.filters
@@ -1,101 +1,33 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <Filter Include="Source Files">
-      <UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
-      <Extensions>cpp;c;cc;cxx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
-    </Filter>
-    <Filter Include="Header Files">
-      <UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
-      <Extensions>h;hh;hpp;hxx;hm;inl;inc;xsd</Extensions>
-    </Filter>
-    <Filter Include="Resource Files">
-      <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
-      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
-    </Filter>
+    <ClCompile Include="Main.cpp" />
+    <ClCompile Include="micromouse\Controller.cpp" />
+    <ClCompile Include="micromouse\FlagMatrix.cpp" />
+    <ClCompile Include="micromouse\Maze.cpp" />
+    <ClCompile Include="micromouse\Motor.cpp" />
+    <ClCompile Include="micromouse\MouseBot.cpp" />
+    <ClCompile Include="micromouse\Node.cpp" />
+    <ClCompile Include="micromouse\Path.cpp" />
+    <ClCompile Include="micromouse\RobotIO.cpp" />
+    <ClCompile Include="micromouse\IRSensor.cpp" />
+    <ClCompile Include="micromouse\Vector.cpp" />
+    <ClCompile Include="micromouse\VirtualMaze.cpp" />
+    <ClCompile Include="micromouse\Magnetometer.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Main.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="micromouse\Controller.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="micromouse\FlagMatrix.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="micromouse\Maze.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="micromouse\MouseBot.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="micromouse\Node.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="micromouse\Path.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="micromouse\Vector.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="micromouse\RobotIO.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="micromouse\VirtualMaze.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="micromouse\IRSensor.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-<<<<<<< HEAD
-    <ClCompile Include="Magnetometer.cpp">
-=======
-    <ClCompile Include="micromouse\Motor.cpp">
->>>>>>> origin/joshua_motor_class
-      <Filter>Source Files</Filter>
-    </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="micromouse\Controller.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="micromouse\FlagMatrix.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="micromouse\Logger.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="micromouse\Maze.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="micromouse\MouseBot.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="micromouse\Node.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="micromouse\Path.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="micromouse\Vector.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="micromouse\RobotIO.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="micromouse\VirtualMaze.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="micromouse\IRSensor.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-<<<<<<< HEAD
-    <ClInclude Include="Magnetometer.h">
-=======
-    <ClInclude Include="micromouse\Motor.h">
->>>>>>> origin/joshua_motor_class
-      <Filter>Header Files</Filter>
-    </ClInclude>
+    <ClInclude Include="micromouse\Controller.h" />
+    <ClInclude Include="micromouse\FlagMatrix.h" />
+    <ClInclude Include="micromouse\Logger.h" />
+    <ClInclude Include="micromouse\Maze.h" />
+    <ClInclude Include="micromouse\Motor.h" />
+    <ClInclude Include="micromouse\MouseBot.h" />
+    <ClInclude Include="micromouse\Node.h" />
+    <ClInclude Include="micromouse\Path.h" />
+    <ClInclude Include="micromouse\RobotIO.h" />
+    <ClInclude Include="micromouse\IRSensor.h" />
+    <ClInclude Include="micromouse\Vector.h" />
+    <ClInclude Include="micromouse\VirtualMaze.h" />
+    <ClInclude Include="micromouse\Magnetometer.h" />
   </ItemGroup>
 </Project>
